+++ Robust.Client/Graphics/Clyde/Clyde.LightRendering.cs
@@ -19,6 +19,7 @@ using Robust.Client.ComponentTrees;
 using Robust.Shared.Graphics;
 using static Robust.Shared.GameObjects.OccluderComponent;
 using Robust.Shared.Utility;
+using TextureWrapMode = Robust.Shared.Graphics.TextureWrapMode;
 using Vector4 = Robust.Shared.Maths.Vector4;
 
 namespace Robust.Client.Graphics.Clyde
