+++ Robust.Client/Graphics/Clyde/Clyde.Textures.cs
@@ -19,6 +19,7 @@ using OGLTextureWrapMode = OpenToolkit.Graphics.OpenGL.TextureWrapMode;
 using PIF = OpenToolkit.Graphics.OpenGL4.PixelInternalFormat;
 using PF = OpenToolkit.Graphics.OpenGL4.PixelFormat;
 using PT = OpenToolkit.Graphics.OpenGL4.PixelType;
+using TextureWrapMode = Robust.Shared.Graphics.TextureWrapMode;
 
 namespace Robust.Client.Graphics.Clyde
 {
@@ -465,15 +466,15 @@ namespace Robust.Client.Graphics.Clyde
         {
             var white = new Image<Rgba32>(1, 1);
             white[0, 0] = new Rgba32(255, 255, 255, 255);
-            _stockTextureWhite = (ClydeTexture) ClydeTexture.LoadFromImage(white);
+            _stockTextureWhite = (ClydeTexture) Texture.LoadFromImage(white);
 
             var black = new Image<Rgba32>(1, 1);
             black[0, 0] = new Rgba32(0, 0, 0, 255);
-            _stockTextureBlack = (ClydeTexture) ClydeTexture.LoadFromImage(black);
+            _stockTextureBlack = (ClydeTexture) Texture.LoadFromImage(black);
 
             var blank = new Image<Rgba32>(1, 1);
             blank[0, 0] = new Rgba32(0, 0, 0, 0);
-            _stockTextureTransparent = (ClydeTexture) ClydeTexture.LoadFromImage(blank);
+            _stockTextureTransparent = (ClydeTexture) Texture.LoadFromImage(blank);
         }
 
         /// <summary>
@@ -602,48 +603,6 @@ namespace Robust.Client.Graphics.Clyde
 
         private sealed class ClydeTexture : OwnedTexture
         {
-            public static Texture Transparent =>
-                IoCManager.Resolve<IClydeInternal>().GetStockTexture(ClydeStockTexture.Transparent);
-
-            public static Texture White =>
-                IoCManager.Resolve<IClydeInternal>().GetStockTexture(ClydeStockTexture.White);
-
-            public static Texture Black =>
-                IoCManager.Resolve<IClydeInternal>().GetStockTexture(ClydeStockTexture.Black);
-
-            /// <summary>
-            ///     Loads a new texture an existing image.
-            /// </summary>
-            /// <param name="image">The image to load.</param>
-            /// <param name="name">The "name" of this texture. This can be referred to later to aid debugging.</param>
-            /// <param name="loadParameters">
-            ///     Parameters that influence the loading of textures.
-            ///     Defaults to <see cref="Robust.Client.Graphics.TextureLoadParameters.Default"/> if <c>null</c>.
-            /// </param>
-            /// <typeparam name="T">The type of pixels of the image. At the moment, images must be <see cref="Rgba32"/>.</typeparam>
-            public static Texture LoadFromImage<T>(Image<T> image, string? name = null,
-                TextureLoadParameters? loadParameters = null) where T : unmanaged, IPixel<T>
-            {
-                var manager = IoCManager.Resolve<IClyde>();
-                return manager.LoadTextureFromImage(image, name, loadParameters);
-            }
-
-            /// <summary>
-            ///     Loads an image from a stream containing PNG data.
-            /// </summary>
-            /// <param name="stream">The stream to load the image from.</param>
-            /// <param name="name">The "name" of this texture. This can be referred to later to aid debugging.</param>
-            /// <param name="loadParameters">
-            ///     Parameters that influence the loading of textures.
-            ///     Defaults to <see cref="Robust.Client.Graphics.TextureLoadParameters.Default"/> if <c>null</c>.
-            /// </param>
-            public static Texture LoadFromPNGStream(Stream stream, string? name = null,
-                TextureLoadParameters? loadParameters = null)
-            {
-                var manager = IoCManager.Resolve<IClyde>();
-                return manager.LoadTextureFromPNGStream(stream, name, loadParameters);
-            }
-
             private readonly Clyde _clyde;
             public readonly bool IsSrgb;
