--- Robust.Client/Graphics/ClydeStockTexture.cs
@@ -1,9 +1,0 @@
-namespace Robust.Client.Graphics
-{
-    internal enum ClydeStockTexture : byte
-    {
-        White,
-        Black,
-        Transparent
-    }
-}
