+++ Robust.Shared/Graphics/ClydeStockTexture.cs
@@ -1,0 +1,9 @@
+namespace Robust.Shared.Graphics
+{
+    public enum ClydeStockTexture : byte
+    {
+        White,
+        Black,
+        Transparent
+    }
+}
