+++ Robust.UnitTesting/Client/Graphics/TextureLoadParametersTest.cs
@@ -1,6 +1,7 @@
 using System.IO;
 using NUnit.Framework;
 using Robust.Client.Graphics;
+using Robust.Shared.Graphics;
 using YamlDotNet.RepresentationModel;
 
 namespace Robust.UnitTesting.Client.Graphics
